Feat. add useMutateDeleteDB
