+++ src/atoms/atoms.js
@@ -2,6 +2,8 @@ import { atom } from "jotai";
 
 export const userAtom = atom("");
 
+export const databasesAtom = atom([]);
+
 export const currentDBIdAtom = atom("");
 export const currentDBNameAtom = atom("");
 export const currentDocIndexAtom = atom(0);
