+++ src/components/Modals/DeleteDatabase/DeleteDBModal.jsx
@@ -1,12 +1,5 @@
 import { useSetAtom, useAtomValue } from "jotai";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-
-import fetchData from "../../../utils/axios";
 import {
-  currentDBIdAtom,
-  currentDBNameAtom,
-  userAtom,
-  currentDocIndexAtom,
   showDeleteDBModalAtom,
   deleteTargetDBIdAtom,
 } from "../../../atoms/atoms";
@@ -16,50 +9,13 @@ import Button from "../../shared/Button";
 import Content from "../SharedItems/Content";
 import ContentWrapper from "../SharedItems/ContentWrapper";
 import Message from "../SharedItems/Message";
-import Loading from "../../shared/Loading";
+import useMutateDeleteDB from "../../../apis/useMutateDeleteDB";
 
-function DeleteDBModal({ databases }) {
-  const queryClient = useQueryClient();
-
-  const { userId } = useAtomValue(userAtom);
+function DeleteDBModal() {
   const deleteTargetDBId = useAtomValue(deleteTargetDBIdAtom);
-
-  const setCurrentDBId = useSetAtom(currentDBIdAtom);
-  const setCurrentDBName = useSetAtom(currentDBNameAtom);
   const setShowDeleteDBModal = useSetAtom(showDeleteDBModalAtom);
-  const setCurrentDocIndex = useSetAtom(currentDocIndexAtom);
 
-  async function deleteDatabase(databaseId) {
-    const response = await fetchData(
-      "DELETE",
-      `/users/${userId}/databases/${databaseId}`,
-    );
-
-    return response.data.databases;
-  }
-
-  const { mutate: fetchDeleteDB, isLoading } = useMutation(deleteDatabase, {
-    onSuccess: result => {
-      if (databases.length === 1) {
-        setCurrentDocIndex(0);
-        setCurrentDBName("");
-      } else {
-        setCurrentDBId(result[0]._id);
-        setCurrentDBName(result[0].name);
-      }
-
-      setShowDeleteDBModal(false);
-      queryClient.refetchQueries(["userDbList"]);
-    },
-    onFailure: () => {
-      console.log("sending user to errorpage");
-      setShowDeleteDBModal(false);
-    },
-  });
-
-  if (isLoading) {
-    return <Loading />;
-  }
+  const fetchDeleteDB = useMutateDeleteDB();
 
   return (
     <Modal onClick={() => setShowDeleteDBModal(false)}>
